requirements.txt
Flask-SQLAlchemy==2.5.1
Flask~=2.0.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
marshmallow==3.12.1
